move asset manager config into macro expansion as much as possible
